fix: 4 bugs do painel administrativo (poção, atributos quebrados, Véu, item some da loja)
- Poção de cura não pode mais carregar efeito_atributo/valor_atributo
  (bug real: Poção de Vida Pequena dando +2 Vitalidade) — migração
  limpa dados existentes e adminItemService passa a zerar esses campos
  sempre que efeito_vida/efeito_mana > 0, tanto criando quanto editando.
- WeaponProperties.valor_bonus_atributo era FLOAT (ArmorProperties
  sempre foi INTEGER) e várias migrations de forja arredondavam pra 1
  casa decimal em vez de inteiro — origem real dos valores quebrados
  (1.1/1.3/1.5). Migração arredondou 344 armas existentes e mudou a
  coluna pra INTEGER; adminItemService arredonda antes de salvar.
- "Véu da Noite"/"Véu do Dia" viram Arma de verdade (Espada/Física/
  Força) — antes eram Acessorio1/Acessorio2 (ArmorProperties) por
  desenho do set Equinócio. Desequipa/cancela Mercado automaticamente
  antes de trocar o tipo, sem abortar (mesmo padrão de
  20261019010000-remove-manoplas-luvas.js).

Testado localmente: migrations rodam limpo do zero, suíte completa
(229 testes) passa.

+++ src/models/WeaponProperties.js
@@ -49,9 +49,14 @@ const WeaponProperties = sequelize.define("WeaponProperties", {
     ),
     allowNull: false,
   },
+  // INTEGER pra combinar com as colunas de bônus equivalentes em
+  // ArmorProperties (bonus_forca/bonus_vitalidade/...) — bug real:
+  // migrations antigas de forja arredondavam pra 1 casa decimal em vez
+  // de inteiro (1.1/1.3/1.5...), e como a coluna era FLOAT o Postgres
+  // deixava passar. Ver migration 20261101020000.
   valor_bonus_atributo: {
-    type: DataTypes.FLOAT,
-    defaultValue: 0.0,
+    type: DataTypes.INTEGER,
+    defaultValue: 0,
     allowNull: false,
   },
 });
